Set up EventHandlers custom type to be modified, but...
I'm not sure if it needs to be a custom type. Could I just store the event
handlers in the controllers folder? I'll have to start looking ahead at
the nature of the event handlers that I'll need to implement, but I'll
probably end up changing the structure of event-handlers.js somehow.