Made some integration tests for mixing tax/subsidy operations; problem:
Yesterday, it looked like when some certain combination of an excise tax,
a subsidy, and a toggling of the graphs was applied, the indicator line
wouldn't be drawn at the right spot, implying that the equilibrium
quantity had become incorrect. This was the reason I made the integration
tests, but I've unexpectedly been unable to reproduce the problem.